refactor: 객체 책임에 따라 패키지 구조 분리
