+++ apps/commerce-api/src/main/java/com/loopers/domain/example/user/service/UserRegisterService.java
@@ -1,5 +1,7 @@
-package com.loopers.domain.example.user;
+package com.loopers.domain.example.user.service;
 
+import com.loopers.application.example.user.RegisterUserCommand;
+import com.loopers.application.example.user.UserResponse;
 import org.springframework.stereotype.Component;
 
 @Component
